Move cloud-init acceptance tests out of definition tests
